# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/ai:formatMarkdown:
        post:
            tags:
                - AIService
            description: |-
                FormatMarkdown restructures plain text into markdown using an instance AI provider,
                 preserving the original text content verbatim.
            operationId: AIService_FormatMarkdown
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FormatMarkdownRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/FormatMarkdownResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:generateFormula:
        post:
            tags:
                - AIService
            description: |-
                GenerateFormula produces a spreadsheet formula from a natural-language prompt
                 using an instance AI provider, returning only the formula string.
            operationId: AIService_GenerateFormula
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenerateFormulaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GenerateFormulaResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:polishText:
        post:
            tags:
                - AIService
            description: |-
                PolishText rewrites a selected span of text using an instance AI provider,
                 following a preset or custom instruction, and returns the rewritten text.
            operationId: AIService_PolishText
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PolishTextRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PolishTextResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/ai:transcribe:
        post:
            tags:
                - AIService
            description: Transcribe transcribes an audio file using an instance AI provider.
            operationId: AIService_Transcribe
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TranscribeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TranscribeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        FormatMarkdownRequest:
            required:
                - text
            type: object
            properties:
                text:
                    type: string
                    description: Required. The plain text to format.
                filename:
                    type: string
                    description: Optional. The source filename, used as context for the model.
        FormatMarkdownResponse:
            type: object
            properties:
                markdown:
                    type: string
                    description: The markdown-formatted text.
...
        GenerateFormulaRequest:
            required:
                - prompt
            type: object
            properties:
                prompt:
                    type: string
                    description: |-
                        Required. A natural-language description of the formula to generate
                         (e.g. "sum the price column", "average of B2 to B10").
                context:
                    type: string
                    description: |-
                        Optional. Spreadsheet context to ground the formula, such as the active
                         cell reference and the sheet's header row / sample rows.
        GenerateFormulaResponse:
            type: object
            properties:
                formula:
                    type: string
                    description: The generated formula, beginning with "=".
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        PolishTextRequest:
            required:
                - text
            type: object
            properties:
                text:
                    type: string
                    description: Required. The selected text to rewrite.
                instruction:
                    type: string
                    description: |-
                        Optional. A custom instruction describing how to rewrite the text
                         (e.g. "make it more formal", "cut the length in half"). When empty,
                         the preset field is used.
                preset:
                    type: string
                    description: |-
                        Optional. A named preset action, used when instruction is empty.
                         One of: "polish", "concise", "expand", "grammar", "tone".
        PolishTextResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The rewritten text.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TranscribeRequest:
            required:
                - audio
            type: object
            properties:
                audio:
                    allOf:
                        - $ref: '#/components/schemas/TranscriptionAudio'
                    description: Required. Audio input.
        TranscribeResponse:
            type: object
            properties:
                text:
                    type: string
                    description: The transcribed text.
        TranscriptionAudio:
            type: object
            properties:
                content:
                    writeOnly: true
                    type: string
                    description: Inline audio bytes.
                    format: bytes
                uri:
                    type: string
                    description: URI for audio content. Reserved for future use.
                filename:
                    type: string
                    description: Optional. The uploaded filename.
                contentType:
                    type: string
                    description: Optional. The MIME type of the input audio.
...